rename demo file
