+++ examples/serve/panels-demo/my_viewer_ext/__init__.py
@@ -3,10 +3,10 @@ from .my_panel_1 import panel as my_panel_1
 
 # from .my_panel_2 import panel as my_panel_2
 
-from .my_panel_4 import panel as my_panel_4
+from .my_panel_3 import panel as my_panel_3
 
 ext = Extension(__name__)
 ext.add(my_panel_1)
 # ext.add(my_panel_2)
 
-ext.add(my_panel_4)
+ext.add(my_panel_3)
