+++ examples/serve/panels-demo/my_viewer_ext/my_panel_3.py
